==> notebooks/synthetic_data/toy_network_mini/p_model/02-building-model-o5.ipynb <==
#### Prerequesites:
* Define schema (make figure) 
* Define structure (make figure)
* Define schema class in python (src/models/model_schemata.py)
* Convert schema + structure into model template (models/templates/p_model.pl)

#### Procedure:
* Load data
* Create lfi script
* Add uncertain evidence to lfi model as probabilistic clauses

%load_ext autoreload
%autoreload 2

# Importing libraries
import pandas as pd
import os

from src.models import model_schemata as schema
from src.models import build_model as build

os.chdir('/Users/magdalena/OneDrive - Queen Mary, University of London/bezzlab/research/projects/phospho_pi/')
# os.chdir('/home/mhuebner/Desktop/bezzlab/research/projects/phospho_pi/')

# reading csv from files into dict
data = {}
data['e_ksea'] = pd.read_csv('data/processed/synthetic_data/toy_network_mini/p_model/e_ksea.csv')
data['p_fc'] = pd.read_csv('data/processed/synthetic_data/toy_network_mini/p_model/p_fc.csv')

# replace e1 with E in enzyme column 
data['e_ksea']['enzyme'] = data['e_ksea']['enzyme'].str.replace('e1', 'E')
data['e_ksea'] = data['e_ksea'].loc[data['e_ksea']['enzyme'] == 'E']
data['e_ksea']

data['p_fc'] = data['p_fc'][data['p_fc']['value'].isin(['inc', 'dec'])]
data['e_ksea'] = data['e_ksea'][data['e_ksea']['value'].isin(['inc', 'dec'])]
data['e_ksea']

# Mapping data to Problog predicates
# make sure to delete 'base' from schemata and AnnotatedDisjunctionClauseTemplate
predicates = {}
predicates['e_ksea'] = schema.EKseaPredicate()
predicates['e_ksea'].add_data(data['e_ksea'], 'enzyme', 'sample', 'value')
predicates['e_ksea'].add_ad_probabilities(data['e_ksea'], 'p_dec', 'p_base', 'p_inc')
predicates['p_fc'] = schema.PFoldChangePredicate()
predicates['p_fc'].add_data(data['p_fc'], 'phosphosite', 'sample', 'value')
predicates['p_fc'].add_ad_probabilities(data['p_fc'], 'p_dec', 'p_base', 'p_inc')

model_lfi = 'models/synthetic_data/toy_network_mini/p_model/p_model_lfi_o6.pl'
for predicate in predicates: 
    ad_generator = build.ProblogStatementGenerator(predicates[predicate])
    clauses = ad_generator.generate_ad_clauses()
    build.insert_statements(model=model_lfi, statements=clauses, location='%% p1::{}'.format(predicate)) # insert into Problog file
