==> notebooks/synthetic_data/toy_network_mini/p_model/03-learning-parameters-o0b.ipynb <==
#### Prerequesites:
* Define schema (make figure) 
* Define structure (make figure)
* Define schema class in python (src/models/model_schemata.py)
* Convert schema + structure into model template (models/templates/p_model.pl)

#### Procedure:
* Load data
* Adjust model script for lfi accordingly
* Create evidence file (if I don't need different evidence files, move to builmodel notebook)
* LFI

1 single instance, 1 pst, 3 enzymes, all samples in one instance, test how many samples can be in one instance, how computing time depends on N p::f

%load_ext autoreload
%autoreload 2

# Importing libraries
import pandas as pd
import numpy as np
import os
import random

from src.models import model_schemata as schema
from src.models import build_model as build
from src.visualization import visualize as viz

# os.chdir('/Users/magdalena/OneDrive - Queen Mary, University of London/bezzlab/research/projects/phospho_pi/')
os.chdir('/home/mhuebner/Desktop/bezzlab/research/projects/phospho_pi/')

#### Training/Testing

# reading csv from files into dict
data = {}
data['e_activity'] = pd.read_csv('data/processed/synthetic_data/toy_network_mini/p_model/e_activity.csv')
data['p_occupancy'] = pd.read_csv('data/processed/synthetic_data/toy_network_mini/p_model/p_occupancy.csv')
data['e_activity']

data['p_occupancy']['prob'] = np.where(data['p_occupancy']['value'] == 'base', 1 - data['p_occupancy']['prob'], data['p_occupancy']['prob'])
data['p_occupancy']['value'] = np.where(data['p_occupancy']['value'] == 'base', np.random.choice(['inc', 'dec'], size=len(data['p_occupancy']['value']), p=[0.5, 0.5]), data['p_occupancy']['value'])
data['e_activity']['prob'] = np.where(data['e_activity']['value'] == 'base', 1 - data['e_activity']['prob'], data['e_activity']['prob'])
data['e_activity']['value'] = np.where(data['e_activity']['value'] == 'base', np.random.choice(['inc', 'dec'], size=len(data['e_activity']['value']), p=[0.5, 0.5]), data['e_activity']['value'])

Splitting data into training/testing

# getting sample names (union of all samples in all datasets)
samples = list(set(data['e_activity']['sample']).union(set(data['p_occupancy']['sample'])))
samples.sort()
# sample x% of samples randomly without replacement with seed
random.seed(613)
train = random.sample(samples, int(len(samples)*0.1))
# train = ['s40', 's23', 's22', 's99', 's7', 's61', 's62', 's63']
# filter data
training_data = {}
training_data['e_activity'] = data['e_activity'][data['e_activity']['sample'].isin(train)].reset_index(drop=True)
training_data['p_occupancy'] = data['p_occupancy'][data['p_occupancy']['sample'].isin(train)].reset_index(drop=True)
train

# Mapping data to Problog predicates
predicates = {}
predicates['e_ksea'] = schema.EKseaPredicate()
predicates['e_ksea'].add_data(training_data['e_activity'], 'enzyme', 'sample', 'value')
predicates['p_fc'] = schema.PFoldChangePredicate()
predicates['p_fc'].add_data(training_data['p_occupancy'], 'phosphosite', 'sample', 'value')

evidence = []
for predicate in predicates:
    evid_generator = build.ProblogStatementGenerator(predicates[predicate])
    evidence = evidence + evid_generator.generate_facts(build.EvidenceTemplate)
evidence[:10]

# write evidence to text file
with open('models/synthetic_data/toy_network_mini/p_model/evidence_o1b.pl', 'w') as f:
    for item in evidence:
        f.write("%s\n" % item)

LFI

DIR = 'models/synthetic_data/toy_network_mini/p_model/'
max_iter = 3
# learning from interpretation
cmd = f'problog lfi {DIR}p_model_lfi_o1b.pl {DIR}evidence_o1b.pl -O {DIR}p_model_o1b.pl -k ddnnf -v -n {max_iter}'
os.system(cmd)
